model.py
"""
Smart Document Finder - High-Accuracy Semantic Search
Uses advanced models and hybrid search for 90%+ accuracy.
"""

from sentence_transformers import SentenceTransformer
from rank_bm25 import BM25Okapi
import pdfplumber
import docx
import re
import numpy as np
import logging
from typing import Dict, List, Tuple, Any
import time

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

CONFIG = {
    "MODEL": "BAAI/bge-small-en-v1.5",  # Best for semantic search
    "CHUNK_SIZE": 500,  # Smaller chunks = better accuracy
    "OVERLAP": 150,     # More overlap for context
    "K": 5,
    "SEMANTIC_WEIGHT": 0.7,  # 70% semantic, 30% BM25
    "BM25_WEIGHT": 0.3,
}

# Load model once
try:
    model = SentenceTransformer(CONFIG["MODEL"])
    logger.info(f"✓ Loaded model: {CONFIG['MODEL']}")
except Exception as e:
    logger.error(f"Failed to load model: {e}")
    raise


def extract_text(file, filename: str) -> str:
    """Extract text from PDF, DOCX, or TXT."""
    try:
        name = filename.lower()
        
        if name.endswith(".pdf"):
            text = ""
            with pdfplumber.open(file) as pdf:
                for page_num, page in enumerate(pdf.pages, 1):
                    extracted = page.extract_text()
                    if extracted:
                        # Add page marker for better tracking
                        text += f"[PAGE {page_num}]\n{extracted}\n"
            return text
            
        elif name.endswith(".docx"):
            doc = docx.Document(file)
            return "\n".join(p.text for p in doc.paragraphs if p.text.strip())
            
        elif name.endswith(".txt"):
            data = file.read()
            return data.decode("utf-8", errors="replace") if isinstance(data, bytes) else str(data)
            
    except Exception as e:
        logger.error(f"Error extracting {filename}: {e}")
        return ""


def is_definition_chunk(text: str, query: str) -> bool:
    """Check if chunk contains a definition or explanation."""
    text_lower = text.lower()
    query_lower = query.lower()
    
    # Definition patterns
    patterns = [
        f"{query_lower} is",
        f"definition of {query_lower}",
        f"what is {query_lower}",
        f"{query_lower} refers to",
        f"{query_lower} means",
        f"{query_lower}:",
    ]
    
    return any(pattern in text_lower for pattern in patterns)


def chunk_text(text: str, chunk_size: int = CONFIG["CHUNK_SIZE"], 
               overlap: int = CONFIG["OVERLAP"]) -> List[str]:
    """Split text into chunks with metadata."""
    if not text or not text.strip():
        return []
    
    # Extract page markers and normalize whitespace
    text_normalized = re.sub(r"\s+", " ", text).strip()
    
    chunks = []
    start = 0
    
    while start < len(text_normalized):
        end = min(start + chunk_size, len(text_normalized))
        
        # Try to break at sentence boundary
        if end < len(text_normalized):
            for pos in range(end, max(start + int(chunk_size * 0.7), end - 50), -1):
                if text_normalized[pos] in '.!?':
                    end = pos + 1
                    break
        
        chunk = text_normalized[start:end].strip()
        if chunk:
            chunks.append(chunk)
        
        start = end - overlap if end < len(text_normalized) else end
    
    logger.info(f"Created {len(chunks)} chunks")
    return chunks


def build_index(chunks: List[str]) -> Dict[str, Any]:
    """Build hybrid index with semantic embeddings and BM25."""
    if not chunks:
        raise ValueError("No chunks to index")
    
    logger.info(f"Building index for {len(chunks)} chunks...")
    start_time = time.time()
    
    # Generate semantic embeddings
    logger.info("Generating embeddings...")
    embeddings = model.encode(chunks, batch_size=32, show_progress_bar=True, convert_to_numpy=True)
    
    # Normalize for cosine similarity
    embeddings_normalized = embeddings / (np.linalg.norm(embeddings, axis=1, keepdims=True) + 1e-10)
    
    # Build BM25 index
    logger.info("Building BM25 index...")
    tokenized_chunks = [chunk.split() for chunk in chunks]
    bm25 = BM25Okapi(tokenized_chunks)
    
    elapsed = time.time() - start_time
    logger.info(f"✓ Index built in {elapsed:.2f}s")
    
    return {
        "embeddings": embeddings_normalized,
        "bm25": bm25,
        "chunks": chunks,
        "time": elapsed
    }


def search(query: str, index: Dict[str, Any], k: int = CONFIG["K"]) -> List[Tuple[str, float]]:
    """
    Hybrid search: 70% semantic + 30% BM25.
    Returns chunks with highest combined score.
    """
    if not query or not query.strip():
        raise ValueError("Query empty")
    
    logger.info(f"Searching: '{query}'")
    start = time.time()
    
    # SEMANTIC SEARCH
    q_emb = model.encode([query])
    q_norm = q_emb / (np.linalg.norm(q_emb, keepdims=True) + 1e-10)
    semantic_scores = (index["embeddings"] @ q_norm.T).ravel()
    
    # BM25 SEARCH
    query_tokens = query.lower().split()
    bm25_scores = index["bm25"].get_scores(query_tokens)
    
    # Normalize BM25 scores to [0, 1]
    if bm25_scores.max() > 0:
        bm25_scores = bm25_scores / bm25_scores.max()
    
    # HYBRID SCORE: weighted combination
    combined_scores = (
        CONFIG["SEMANTIC_WEIGHT"] * semantic_scores + 
        CONFIG["BM25_WEIGHT"] * bm25_scores
    )
    
    # Get top results
    top_k_idx = np.argsort(combined_scores)[::-1][:k*3]  # Get 3x to filter
    
    # Filter and rank: prioritize definition chunks
    results = []
    for idx in top_k_idx:
        chunk = index["chunks"][idx]
        score = combined_scores[idx]
        
        # Boost score if it's a definition
        if is_definition_chunk(chunk, query):
            score = min(score * 1.3, 1.0)  # Boost but cap at 1.0
        
        if score > 0.1:  # Minimum threshold
            results.append((chunk, float(score)))
    
    # Sort by score and take top k
    results = sorted(results, key=lambda x: x[1], reverse=True)[:k]
    
    logger.info(f"Found {len(results)} results in {time.time()-start:.3f}s")
    logger.info(f"Scores: {[f'{s:.2f}' for _, s in results]}")
    
    return results


def estimate_page(chunk_text: str, full_text: str) -> int:
    """Estimate page number from chunk position."""
    try:
        # Extract page markers
        page_matches = re.finditer(r'\[PAGE (\d+)\]', full_text)
        pages = [(m.start(), int(m.group(1))) for m in page_matches]
        
        if not pages:
            return 1
        
        pos = full_text.find(chunk_text)
        if pos < 0:
            return 1
        
        # Find which page this position falls in
        for i in range(len(pages) - 1):
            if pages[i][0] <= pos < pages[i+1][0]:
                return pages[i][1]
        
        return pages[-1][1] if pages else 1
    except:
        return 1
