requirements.txt
# Smart Document Finder - Dependencies
# Python 3.8+

# Core
streamlit==1.28.1

# Better embeddings model (state-of-the-art for semantic search)
# BAAI/bge-small-en-v1.5 is faster and more accurate than all-MiniLM
sentence-transformers==2.2.2
torch>=2.0.0

# Data processing
numpy==1.24.3
scipy==1.11.2

# Document processing
pdfplumber==0.10.3
python-docx==0.8.11

# Optional: Better search with re-ranking
rank-bm25==0.2.2

# Optional GPU (comment out if no GPU)
# torch-cuda support would go here

# Utilities
scikit-learn==1.3.0
python-dotenv==1.0.0

